adjust program

+++ gpt2/cuda/softmax.cuh
@@ -1,5 +1,6 @@
 #pragma once
+#include <cstddef>
 namespace gpt2cuda {
-    void BatchSoftmaxForward(float* outputs , float * const  inputs,int B,int T,int V,int Vp);
-    void BatchSoftmaxBackward(float* d_inputs, float * const d_outputs,float const * inputs , int B ,int T ,int V,int Vp);
+    void BatchSoftmaxForward(float* outputs , float * const  inputs,size_t B,size_t T,size_t V,size_t Vp);
+    void BatchSoftmaxBackward(float* d_inputs, float * const d_outputs,float const * inputs , size_t B ,size_t T ,size_t V,size_t Vp);
 }
